DFTDP# 279 Move letters-to-driver to shared library -Move letters to driver component to shared library - Added the component to public api.ts - Added api/models folder to shared library - Updated driver portal letters to driver componet to use the shared component
