+++ driver-portal/src/ClientApp/src/app/letters-to-driver/letters-to-driver.component.html
@@ -1,169 +1,1 @@
-<div class="container">
-  <div class="list ">
-    <h2>Letters To Driver</h2>
-  <div class="submissions">
-    <div>
-      <p>
-        The Letters to Driver tab displays all the letters that RSBC has sent to
-        you about this case. You can download these letters.
-      </p>
-      <p>
-        Find out more about how to report if a letter is not addressed to you or
-        how to submit an appeal
-      </p>
-    </div>
-    <div>
-      <app-quick-links></app-quick-links>
-    </div>
-  </div>
-
-  <div class="list">
-    <div>
-      <mat-card
-        *ngFor="let letter of filteredDocuments"
-        [ngClass]="{
-          open: letter.documentId && isExpanded[letter.documentId]
-        }"
-      >
-        <mat-card-content class="closedCaseContent">
-          <div style="text-align: center">
-            <mat-icon
-              (click)="toggleIsExpandable(letter.documentId)"
-              class="expandedIcon"
-            >
-              expand_circle_down
-            </mat-icon>
-          </div>
-          <div style="display: flex; flex-direction: column">
-            <div class="case-details-border">&nbsp;</div>
-            <div>&nbsp;</div>
-          </div>
-          <div style="flex: 1">
-            <div class="caseDetailRow">
-              <div>
-                <label>Created Date </label>
-                <div>
-                  <b>{{ letter.createDate | date : "longDate" }}</b>
-                </div>
-              </div>
-
-              <div>
-                <label>Letter Topic</label>
-                <div>
-                  <b>
-                    {{ letter.documentType }}
-                    @if(letter?.documentSubType) { -
-                    {{ letter?.documentSubType }}
-                    }
-                  </b>
-                </div>
-              </div>
-              <div>
-                <div>
-                  <button
-                    mat-flat-button
-                    class="primaryButton"
-                    (click)="downloadLetters(letter?.documentId)"
-                  >
-                    Download
-                  </button>
-                </div>
-              </div>
-            </div>
-
-            <div
-              class="caseDetailRow"
-              *ngIf="letter.documentId && isExpanded?.[letter.documentId]"
-              style="
-                border-top: 2px solid gray;
-                padding-top: 20px;
-                padding-bottom: 20px;
-              "
-            >
-              <div>
-                <label>Case Type</label>
-                <div>
-                  <b>{{ letter.caseType }}</b>
-                </div>
-              </div>
-              <div>
-                <label>Case Id</label>
-                <div>
-                  <b>{{ letter.idCode }}</b>
-                </div>
-              </div>
-              <div>
-                <label>Description</label>
-                <div>
-                  <b>{{ letter.description }}</b>
-                </div>
-              </div>
-            </div>
-          </div>
-        </mat-card-content>
-      </mat-card>
-    </div>
-    <div *ngIf="!isLoading && letterDocuments?.length === 0">
-      <mat-card>
-        <mat-card-content>
-          <b>No Letters found for this case</b>
-        </mat-card-content>
-      </mat-card>
-    </div>
-  </div>
-
-  <div
-    class="numberOfItems"
-    *ngIf="
-    letterOutDocuments &&
-    letterOutDocuments.length > 0 &&
-      filteredDocuments &&
-      filteredDocuments.length > 0
-    "
-  >
-    Showing {{ filteredDocuments.length }} of
-    {{ letterOutDocuments.length }} records
-    <br />
-    <a
-      *ngIf="filteredDocuments.length < letterOutDocuments.length"
-      (click)="viewMore()"
-      class="viewMore"
-      ><b style="color: #003366">View more records</b>
-      <mat-icon class="ViewMoreIcon">expand_circle_down</mat-icon></a
-    >
-  </div>
-
-  <div class="list" id="definitions">
-    <h2>Definitions</h2>
-    <div class="accordion-btn-container">
-      <a class="accordion-btn show-btn" (click)="accordion.openAll()"
-        >Expand All</a
-      >
-      <span class="accordion-bar">&nbsp;|&nbsp;</span>
-      <a class="accordion-btn hide-btn" (click)="accordion.closeAll()"
-        >Collapse All</a
-      >
-    </div>
-    <div style="padding-bottom: 30px">
-      <mat-accordion multi>
-        <mat-expansion-panel>
-          <mat-expansion-panel-header class="panel-header">
-            <mat-panel-title class="accordian-Title">
-              Case Type
-            </mat-panel-title>
-          </mat-expansion-panel-header>
-          <app-case-type></app-case-type>
-        </mat-expansion-panel>
-        <mat-expansion-panel>
-          <mat-expansion-panel-header class="panel-header">
-            <mat-panel-title class="accordian-Title">
-              Letter Topic</mat-panel-title
-            >
-          </mat-expansion-panel-header>
-          <app-letter-topic></app-letter-topic>
-        </mat-expansion-panel>
-      </mat-accordion>
-    </div>
-  </div>
-  </div>
-</div>
+<app-shared-letters-to-driver [caseManagementService]="caseManagementService" ></app-shared-letters-to-driver>
