+++ shared-portal-ui/projects/core-ui/src/lib/api/index.ts
@@ -1,0 +1,1 @@
+export * from './models';
