+++ shared-portal-ui/projects/core-ui/src/lib/api/models/document.ts
@@ -1,0 +1,25 @@
+/* tslint:disable */
+/* eslint-disable */
+import { Driver } from './driver';
+export interface Document {
+  bcMailSent?: boolean | null;
+  businessArea?: string | null;
+  caseId?: string | null;
+  caseType?: string | null;
+  createDate?: string;
+  description?: string | null;
+  documentId?: string | null;
+  documentSubType?: string | null;
+  documentType?: string | null;
+  documentTypeCode?: string | null;
+  documentUrl?: string | null;
+  driver?: Driver;
+  dueDate?: string | null;
+  faxReceivedDate?: string | null;
+  fileContents?: string | null;
+  idCode?: string | null;
+  importDate?: string | null;
+  sequenceNumber?: number | null;
+  submittalStatus?: string | null;
+  userId?: string | null;
+}
