+++ shared-portal-ui/projects/core-ui/src/lib/api/models/driver.ts
@@ -1,0 +1,9 @@
+/* tslint:disable */
+/* eslint-disable */
+export interface Driver {
+  flag51?: boolean | null;
+  lastName?: string | null;
+  licenseNumber?: string | null;
+  loadedFromICBC?: boolean | null;
+  medicalIssueDate?: string | null;
+}
