+++ shared-portal-ui/projects/core-ui/src/lib/api/models/index.ts
@@ -1,0 +1,2 @@
+export * from "./document";
+export * from "./driver";
