+++ shared-portal-ui/projects/core-ui/src/lib/letters-to-driver/letters-to-driver.component.scss
@@ -1,0 +1,86 @@
+.submissions {
+  display: flex;
+  flex-direction: row;
+  padding-top: 24px;
+  gap: 100px;
+}
+.submissionRequirement {
+  display: flex;
+  width: 100%;
+  justify-content: space-between;
+}
+
+.mat-mdc-card {
+  border: 1px solid gray;
+  box-shadow: none !important;
+  margin: 30px 0 30px 0;
+  padding: 32px;
+  margin-bottom: 24px;
+
+  .mat-mdc-card-header {
+    padding-top: 12px;
+
+    .mat-mdc-card-header-text {
+      width: 100%;
+      vertical-align: middle;
+    }
+  }
+}
+
+
+
+.closedCaseContent {
+  display: flex;
+  flex-direction: row;
+  gap: 24px;
+}
+
+.case-details-border {
+  border-left: 2px solid #808080;
+  height: 45%;
+}
+
+.caseDetailRow {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  line-height: 32px;
+  padding-bottom: 24px;
+  align-content: flex-start;
+}
+
+.caseDetailRow > div {
+  flex-basis: 220px;
+}
+
+.expandedIcon {
+  transform: scale(1.5);
+  color: #003366;
+  width: 50px;
+}
+
+.open {
+  background-color: #f1f1f3;
+  border: 1px solid #003366;
+  cursor: pointer;
+}
+
+@media only screen and (max-width: 896px) {
+  .submissions {
+    display: block;
+  }
+
+  .closedCaseContent {
+    flex-direction: column-reverse;
+  }
+
+  .caseDetailRow {
+    gap: 50px;
+  }
+
+  .case-details-border {
+    border-left: none;
+    border-top: 4px solid #808080;
+    height: 8px;
+  }
+}
