+++ shared-portal-ui/projects/core-ui/src/lib/letters-to-driver/letters-to-driver.component.spec.ts
@@ -1,0 +1,17 @@
+describe('LettersToDriverComponent', () => {
+  // let component: LettersToDriverComponent;
+  // let fixture: ComponentFixture<LettersToDriverComponent>;
+
+  // beforeEach(() => {
+  //   TestBed.configureTestingModule({
+  //     declarations: [LettersToDriverComponent]
+  //   });
+  //   fixture = TestBed.createComponent(LettersToDriverComponent);
+  //   component = fixture.componentInstance;
+  //   fixture.detectChanges();
+  // });
+
+  it('should create', () => {
+    // expect(component).toBeTruthy();
+  });
+});
